# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/bitcoin": {
            "put": {
                "description": "Update an Bitcoin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bitcoin"
                ],
                "summary": "Update an Bitcoin",
                "parameters": [
                    {
                        "description": "Update Bitcoin Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateBitcoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateBitcoinResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_bitcoin.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_bitcoin.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_bitcoin.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a Bitcoin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bitcoin"
                ],
                "summary": "Create a Bitcoin",
                "parameters": [
                    {
                        "description": "Create Bitcoin Params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateBitcoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.CreateBitcoinParams"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_bitcoin.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_bitcoin.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/bitcoin/{id}": {
            "get": {
                "description": "Show an Bitcoin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bitcoin"
                ],
                "summary": "Show Bitcoin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Show Bitcoin Request",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShowBitcoinResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_bitcoin.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_bitcoin.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Bitcoin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bitcoin"
                ],
                "summary": "Delete a Bitcoin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete Bitcoin Param",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteBitcoinResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_bitcoin.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_bitcoin.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/fixed-expense/{id}": {
            "get": {
                "description": "Show an FixedExpense",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fixed expense"
                ],
                "summary": "Show FixedExpense",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Show FixedExpense Request",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShowFixedExpenseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_fixed_expense.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_fixed_expense.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an FixedExpense",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fixed expense"
                ],
                "summary": "Delete an FixedExpense",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete FixedExpense Param",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteFixedExpenseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_fixed_expense.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_fixed_expense.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/investment": {
            "put": {
                "description": "Update an Investment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investment"
                ],
                "summary": "Update an Investment",
                "parameters": [
                    {
                        "description": "Update Investment Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateInvestmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateInvestmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_investment.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_investment.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_investment.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create an Investment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investment"
                ],
                "summary": "Create an Investment",
                "parameters": [
                    {
                        "description": "Create Investment Params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateInvestmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateInvestmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_investment.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_investment.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/investment/{id}": {
            "get": {
                "description": "Show an Investment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investment"
                ],
                "summary": "Show Investment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Show Investment Request",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShowInvestmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_investment.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_investment.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an Investment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "investment"
                ],
                "summary": "Delete an Investment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete Investment Param",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteInvestmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_investment.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_isaiaspereira307_gowallet_handlers_investment.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.Bitcoin": {
            "type": "object",
            "properties": {
                "bank_account_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "purchase_date": {
                    "type": "string"
                },
                "purchase_price": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                }
            }
        },
        "db.CreateBitcoinParams": {
            "type": "object",
            "properties": {
                "bankAccountID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "purchaseDate": {
                    "type": "string"
                },
                "purchasePrice": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                }
            }
        },
        "db.FixedExpense": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "bank_account_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "db.Investment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "bank_account_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_isaiaspereira307_gowallet_handlers_bitcoin.ErrorResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_isaiaspereira307_gowallet_handlers_fixed_expense.ErrorResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_isaiaspereira307_gowallet_handlers_investment.ErrorResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CreateBitcoinRequest": {
            "type": "object",
            "required": [
                "bank_account_id"
            ],
            "properties": {
                "bank_account_id": {
                    "type": "integer"
                },
                "purchase_date": {
                    "type": "string"
                },
                "purchase_price": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CreateInvestmentRequest": {
            "type": "object",
            "required": [
                "amount",
                "bank_account_id"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "bank_account_id": {
                    "type": "integer"
                }
            }
        },
        "handlers.CreateInvestmentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Investment"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.DeleteBitcoinResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Bitcoin"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.DeleteFixedExpenseResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.FixedExpense"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.DeleteInvestmentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Investment"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ShowBitcoinResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Bitcoin"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.ShowFixedExpenseResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.FixedExpense"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.ShowInvestmentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Investment"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdateBitcoinRequest": {
            "type": "object",
            "required": [
                "bank_account_id",
                "id",
                "purchase_date",
                "purchase_price",
                "quantity"
            ],
            "properties": {
                "bank_account_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "purchase_date": {
                    "type": "string"
                },
                "purchase_price": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                }
            }
        },
        "handlers.UpdateBitcoinResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Bitcoin"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdateInvestmentRequest": {
            "type": "object",
            "required": [
                "amount",
                "id"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "handlers.UpdateInvestmentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/db.Investment"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}