# handlers/fixed_expense/getFixedExpense.go
package handlers

import (
	"math"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
)

// @BasePath /api/v1
// @Summary Show FixedExpense
// @Description Show an FixedExpense
// @Tags fixed expense
// @Accept json
// @Produce json
// @Param id path string true "Show FixedExpense Request"
// @Success 200 {object} ShowFixedExpenseResponse
// @Failure 400 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Router /fixed-expense/{id} [get]
func GetFixedExpense(ctx *gin.Context) {
	id := ctx.Param("id")
	idInt64, err := strconv.ParseInt(id, 10, 64)
	if err != nil || idInt64 > math.MaxInt32 || idInt64 < math.MinInt32 {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "invalid id"})
		return
	}
	idInt32 := int32(idInt64)
	fixedExpense, err := queries.GetFixedExpenseById(ctx, idInt32)
	if err != nil {
		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "failed to get fixed expense"})
		return
	}

	ctx.JSON(http.StatusOK, fixedExpense)
}